ovl: verify origin of merge dir lower
When inodes index feature is enabled, verify that lower merge dir found
by name matches the origin file handle stored in xattr on upper dir.
If origin xattr does not exist, update it with the lower found by name.

If lower dir does not match the origin fh stored in upper dir, or if no
lower dir was found by name, try to decode the stored origin fh and
use the result as the merge dir lower dentry.

This change is needed for indexing of merge directories. A merge
directory is indexed by the file handle of the uppermost lower dir and
the index will have a reference to the upper dir by file handle.
Following down from upper by origin file handle ensures the integrity
of the index for non-stale origin. index entries with stale origin
are cleaned up on mount anyway.

This allows an overlayfs to cope better with changes to lower layer,
like rename and remove of lower merge dir.

A lower merge directory followed by origin fh cannot be followed down
again by name, so practically, overlayfs is not able to cope well with
lower layer changes when there is more than a single lower layer.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>