ovl: copy on read with consistent_fd feature
Overlayfs copies up a file when the file is opened for write.  Writes
to the returned file descriptor are not visible to file descriptors
that were opened for read prior to copy up.

If this config option is enabled then overlay filesystems will default
to copy up a file that is opened for read to avoid this inconsistency.
In this case, it is still possible to turn off consistent fd globally
with the "consistent_fd=off" module option or on a filesystem instance
basis with the "consistent_fd=off" mount option.

The feature will not be turned on for an overlay mount unless all
the layers are on the same underlying filesystem and this filesystem
supports clone.

On ro mount there are only ro fd's, so all fd's of the same file
are always consistent on a ro overlay mount.  But, if overlay can
be later remounted rw, we need to copy on read anyway, so that
ro fd that was opened during ro mount will be consistent with
rw fd that is opened after remount rw.

Therefore, we disable consistent_fd feature only in case there are
upper/work dirs and overlay cannot be remounted rw.

The copy on read behavior introduces a slight change in d_real()
semantics. Now d_real() may return an error with NULL inode argument
also in the zero open flags case. vfs API documentation has been updated.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>