ovl: encode/decode pure upper directory file handles
Implement get_parent() operation to reconnect a pure upper overlay
parent directory. To allow reconnecting of pure upper directory,
based on its real parent, we restrict the implementation to
encoding of directories with pure upper ancestry up to overlay root.

This fixes a variant of xfstest test generic/426 which creates a
directory, encodes a file handle, drops caches and decodes it.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>