APP-14660 Fix crash after changing sync mode to Classic
When changing sync mode to Classic, the .ctera/cloudgateway
bind mount is unmounted and the mnt pointer in the fanotify
root watch that was pointing to that mount becomes stale.

While the watch itself should be removed when changing to Classic
mode, there could still be events on-the-fly referring to the
stale mnt pointer.

APP-14705 reports about the kernel hitting the warning
WARN_ON(dentry->d_sb != mnt->mnt_sb) in fanotify_should_send_event(),
which is the case where the stale mnt pointer is detected on time
before making any damage.
APP-14660 reports about the kernel trying to mntput the stale mnt
pointer and crashing.

To solve APP-14660, we don't store the mnt pointer at all after
filtering the event in fanotify_should_send_event(). storing the
mnt pointer is only needed for reporting an open file descriptor
with the change notification, but our implementation for Linux
of FileChangeListenerLinuxEvents ignores those file descriptors
anyway.

This change does NOT fix APP-14705.

Signed-off-by: Amir Goldstein <amir@ctera.com>