ovl: encode non pure upper file handles
Encode a non pure upper overlay inode as struct ovl_fh encoding
of the copy up origin inode. The encoded real inode, is also the
inode used as the key to hash the overlay inode.

When we decode a file handle obtained from a lower dentry, that
lower may have already been copied up. If that lower is not a
hardlink, then it may not be indexed on copy up. In that case,
the way we get to the upper inode when decoding is by looking up
the overlay using the same path as lower dentry, but to do that,
we need a connected lower dentry, so we encode a connectable lower
file handle even if we were asked for a non-connectable file
handle.

The problem with connectable file handles is that they may not be
unique, because they contain the parent inode encoding and dentry
can change its parent. We rely on the assumption that lower layer
is not expected to be changed and therefore the connectable lower
file handle is "likely unique". If lower layer is changed, then the
new encoded file handle may appear to nfsd as a new file and the old
file handle may become stale.

We cannot assume uniqeness of connectable file handle for a lower
hardlink, so instead, we copy up the lower hardlink before encoding
the file handle, so we can use the index on decode.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>